Add position to tables in the AdminServiceStore for listed items
The order of members, services, arguments, etc must be consistent
across splinter nodes, do to the hashing that is a part of circuit
management. When the storage was a yaml file, this was kept
consistent.

However, with the change to use diesel backends the order was switched
to be alphabetical. This caused an issue with working with 0.4 splinter
nodes.

This commit updates the tables of items that are listed (requiring a
separate table) to include a position column, so proposals and
circuits can be rebuild correctly.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>