Add `AdminServiceEventStore` to `StoreFactory`
Adds a `get_admin_service_event_store` method to the `StoreFactory`
trait and implementations for the memory-, SQLite-, and
PostgreSQL-backends.

Also renames the `DieselAdminServiceEventStore` constructor method from
`_new` to `new`, as this function is now being used within the
`StoreFactory` module and no longer needs the underscore prefix to avoid
compiler warnings.

Signed-off-by: Shannyn Telander <telander@bitwise.io>