Add RoleBasedAuthorizationHandler
This change adds an AuthorizationHandler implementation backed by the
RoleBasedAuthorizationStore.  It determines if a Key or User identity is
allowed based on its assigned roles and the permissions those roles
define. If an identity doesn't have an roles or is not assigned a role
with the requested permission, the handler defers to the next handler in
the chain (via a Continue).

This handler is guarded by the feature "authorization-handler-rbac".

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>