Add actix 1 paging extraction method
This change introduces a function that will extract paging information
from a query and convert it into a PagingQuery struct that contains the
limit and offset. Any errors are returned in a single BadPagingRequest
error that can be converted into an HttpResponse future containing a
BadRequest.

This change also includes updating existing routes that make use of
these functions.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>