Add user_id to Profile struct
Add a user_id field to the Profile struct to hold the Biome ID for the
user that the profile belongs to. Remove NewProfileModel and add
user_id to the ProfileModel struct. Add user_id to the user_profile
table and update the table to use user_id as the primary key. Update
sqlite profile store tests accordingly.

Signed-off-by: Isabel Tomb <tomb@bitwise.io>