Add conversion methods for admin messages
Adds an `into_circuit_proto` used by the `CreateCircuitBuilder` to turn
the `CreateCircuit` into the `admin::proto::Circuit`. This is necessary
as the `into_proto` method turns the `CreateCircuitBuilder` into a
`CircuitCreateRequest`, rather than the internal `admin::proto::Circuit`
struct. This is necessary to implement the `CircuitDisbandRequest`,
without wrapping the circuit in a `CircuitCreateRequest`.

Also adds conversions for the native messages' enums, which allows the
`AdminServiceStore`'s definitions of the enums to be correctly converted
to the `admin::messages` version of these structs.

Signed-off-by: Shannyn Telander <telander@bitwise.io>