Update `circuit_status` column type and default
Updates the migration files adding the `circuit_status` column to the
database to have a default value of 1, equivalent to
`CircuitStatus::Active`. Also updates the postgres migration file to
change the type from `INTEGER` to `SMALLINT` so the `circuit_status`
value is able to be converted accurately from a postgres backend.

Signed-off-by: Shannyn Telander <telander@bitwise.io>