Implement `CircuitDisbandRequest` in admin service
Implements the handling of a `CircuitDisbandRequest`, including payload
validation. The `CircuitDisbandRequest` is sent between admin services
to agree on disbanding a circuit, meaning the circuit is only available
locally. When disbanded, the circuit's associated services are shutdown
and the circuit is removed from the admin service's routing table, but
the circuit is still available locally to be viewed in it's `Disbanded`
state.

This implementation is behind the experimental `circuit-disband`
feature.

Signed-off-by: Shannyn Telander <telander@bitwise.io>