Allow for 0.4 compatible circuits to be submitted
Changes how the admin service verifies what to do with a proposal before
committing it to state. Before, the action was determined by the circuit
proposal's `circuit_status`, which will not be set for version 1, or 0.4
compatible, circuits. This value should default to `Active` when being
stored, but otherwise should not affect the admin service.

This specifically changes the match statement to match a `Disbanded`
status, which is valid for the `CircuitDisbandRequest`, but will
go to upgrading the proposal to a circuit if the proposal's
`circuit_status` is `ACTIVE` or unset, to allow for 0.4 compatibility
when the `circuit_status` field will not be set. The `CircuitDisbandRequest`
functionality is not supported by 0.4.

Signed-off-by: Shannyn Telander <telander@bitwise.io>