More accurately display proposals
Adds a `PROPOSAL_TYPE` column to the output of the `splinter circuit
proposals` command, to display the `CircuitProposal`'s proposal type.
This enables users to quickly see the intention of the proposal.

Also adds a check before displaying a proposal for the proposal type, to
more accurately reflect the proposal's intention in the introductory
sentence.

Also fixes a typo, `submited` -> `submitted`.

Signed-off-by: Shannyn Telander <telander@bitwise.io>