Fix assignment issue with deleted role
This change fixes an issue where when a role is deleted, and there
exists an assignment with a single reference to the role, the role could
no longer be loaded, causing a 500 server error down stream.  The root
cause was using the builder to create the assignment from the database.
This builder has a requirement that the roles list cannot be empty, but
that constraint doesn't apply in this direction, due to deletions.

Building the assignment directly corrected the issue.  A test has been
added to avoid any regressions.

An alternative solution would be to not allow assigned roles to be
deleted, but this would require a full audit of assignments.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>