Ensure no new peer refs are added when disbanding
Changes the order in the `handle_proposed_circuit` function, which is
used by nodes when receiving a circuit proposal to only add peer refs
if the associated payload has a `CircuitCreateRequest`. Peer refs are
assumed to already exist if the proposal is to disband a circuit, so no
new peer refs should be added.

Signed-off-by: Shannyn Telander <telander@bitwise.io>