Implement `CircuitPurgeRequest` in admin service
Implements the functionality to purge a circuit via the local
`CircuitPurgeRequest`. If a circuit has been disbanded, it is only
available locally and is able to be purged.

This change also adds tests for the `CircuitPurgeRequest` handling
within the admin service. Also updates test helper function that creates
an admin::store::Circuit to include an argument for the
`circuit_status`, which is simpler than having multiple helper
functions and this method better suits the tests added.

This functionality is behind the experimental `circuit-purge` feature,
which is dependent on the `circuit-disband` feature, as circuits must
first be disbanded before they are able to be purged.

Signed-off-by: Shannyn Telander <telander@bitwise.io>