Return error on add/rm of same perm
This change returns an error if the user has tried to add and remove the
same permission within the same command execution. These errors can be
ignored by use of the --force flag.

Additionally, the order at which removals and additions are processed is
changed such that removals are applied only to existing permissions and
additions are then applied after.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>