Implement `Clone` for `AdminServiceEvent`
Adds the `Clone` trait to the `AdminServiceEvent` and `EventType`
structs. The `EventType` is used within the `AdminServiceEvent` and
implementing `Clone` for this struct is necessary to implement `Clone`
on the event struct itself. This allows for greater flexibility in the
memory implementation of the `AdminServiceEventStore` when converting
messages::AdminServiceEvent to the events::AdminServiceEvent and still
being able to return the `AdminServiceEvent` constructed.

Signed-off-by: Shannyn Telander <telander@bitwise.io>