Add event methods to `AdminServiceStore` trait
Adds the `add_event`, `list_events_since`, and
`list_events_by_management_type_since` methods previously implemented
for the `AdminServiceEventStore` to the `AdminServiceStore` trait. Also
adds the `event` module to the admin store to hold the `AdminServiceEvent`
definition used by the admin store.

This change also adds the unimplemented methods to the diesel-backed and
YAML-backed `AdminServiceStore`.

As the `AdminServiceEventStore` is integrated into the
`AdminServiceStore`, the event methods and implementations are behind
the experimental `admin-service-event-store` feature.

Signed-off-by: Shannyn Telander <telander@bitwise.io>