Add `remove_db_files` method to ScabbardState
Adds a method to the `ScabbardState` struct to remove the state and
transaction receipt store LMDB files. This is used within `Scabbard`'s
`destroy` method, which previously only verified the consensus thread
has been shutdown. This change adds an additional step to this method to
remove the state's LMDB files if the consensus shutdown.

Signed-off-by: Shannyn Telander <telander@bitwise.io>