Add `destroy_service` to Orchestrator
Adds a `destroy_service` method used to call the `destroy` method on a
service. Also renames `shutdown_service` to `stop_service` to better
describe it's function, as the `destroy` method call has now been moved
to the `destroy_service` method. This allows for the admin service to
stop a service and then destroy the service, including removing the
state LMDB files, at a later point.

This change also alters the admin service's `shutdown_services` method
to `stop_services` to better represent the functionality, which now just
uses the Orchestrator's `stop_service` method.

Signed-off-by: Shannyn Telander <telander@bitwise.io>