Update RegistryError to use top level errors
Replaces the existing error types with InternalError,
InvalidState, ConstraintViolationError, and
ResourceTemporarilyUnavailableError.

Also adds a error that can be used in the REST API routes.
The Registry error can no longer be used in a web::block,
because Box<dyn Error> is guaranteed to be Send. The
new error will conver the RegistryError to either InvalidState,
which is required for returning a BadRequest, and
InternalError that will be logged and InternalServerError
is returned.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>