Update update_node not to fail if dup endpoint belongs to node
While updating a node, there is a check to see if their is a duplicate
endpoint that already belongs to another node. However, if this endpoint
belongs to the node that is being updated, this should not be counted
as an error. All old data for the node is removed and readded
during the update step.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>