Move allow_keys file to config dir
Moves the backing file for the AllowKeysAuthorizationHandler from the
state directory to the config directory, since it's considered a config
file.

This change requires adding the `config_dir` field to the
`SplinterDaemon` and its builder. Initially, this field is guarded by
the allow keys handler's feature since that is the only code that uses
it.

Signed-off-by: Logan Seeley <seeley@bitwise.io>