Add new memory features that depends on sqlite
Many memory implementation of stores are sqlite
stores using :memory: as the connection string. Adding
this feature allows the crate to enforce that sqlite
will be available when using the memory implementations
from the StoreFactory and the unimplemented!()
implementation used if sqlite was not available can
be removed.

The memory feature is added as stable in libsplinter
and pulled in by splinterd.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>