Convert identity_type DB enum type
This change converts the data type on the columns for identity_type to
an enum value.  In sqlite, this is a string with a check. In postgres,
this is an actual enum type.  Both of these require a bit of tricky
transformations when dealing with custom types, but the result is both a
user-friendly queryable value (a readable string in the column vs an
int) and strongly-typed values to and from the database.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>