Use admin store in place of the admin event store
Changes the type of the `admin_service_event_store` argument used to
construct the admin service to allow for the different configurations of
the admin store and admin event store. Mainly, this allows for
a differently configured event store when the `yaml` storage option is used
for the admin store, as the event store operations are not implemented for
the YAML admin store. Keeping the `event_store` argument allows
splinterd to pass in the `AdminServiceStore` from the store factory when
the YAML admin store is used.

Signed-off-by: Shannyn Telander <telander@bitwise.io>