Add threading::shutdown module
A new ShutdownHandle trait is defined here. Anything which runs a thread
will need to implement ShutdownHandle in the future, which will
standardize libsplinter's approach to shutdown (a.k.a. joining threads).

A shutdown function is defined which may be used to initiate and handle
shutdown.

This approach differs from some of the current code in libsplinter which
defines separate traits for signaling shutdown and waiting for shutdown.
The primary reason for joining them together in ShutdownHandle is so
that code like that in the shutdown() function can be written. This
places some new constraints on the implementations, which must now
provide both on a single struct.

Signed-off-by: Shawn T. Amundson <amundson@bitwise.io>