Remove enable_biome option for splinterd
Removes the `enable_biome` option for the Splinter daemon. The inclusion
of Biome is now a compile-time choice rather than run-time. Previously,
the `enable_biome` option was necessary to be able to run splinterd
without a database, but soon a database will always be required for
splinterd (that change will be made in a separate commit).

For backwards compatibility, the `--enable-biome` CLI flag is kept but
marked as hidden.

Signed-off-by: Logan Seeley <seeley@bitwise.io>