Add initial Actix Web 3 framework support
This adds the start of pieces we need to easily standup and manage the
REST API without any wrapper code (such as the approach taken with the
Actix Web 1 support). This includes support for a RestApi struct and
its lifecycle (RestApiBuilder -> RunnbleRestApi -> RestApi -> shutdown).

This is only an initial cut at the core pieces; much remains like
support for authorization, etc.

Signed-off-by: Shawn T. Amundson <amundson@bitwise.io>