Refactor splinterd database feature
Replaces splinterd's `database` feature with separate features for
PostgreSQL and SQLite support. A compiler error will occur if neither of
the database backend features is enabled. Because one of the database
features will always be required now, this commit also removes the
feature guards around the database URL configuration and makes that
configuration option required.

Signed-off-by: Logan Seeley <seeley@bitwise.io>