Move Biome key routes to new resource provider
Splits the Biome key management routes out into a new
`BiomeKeyManagementRestResourceProvider. This is part of an effort to
make each of the Biome features self-contained.

This includes a couple minor tweaks to support this move:

* Replace the use of the `get_authorized_user` function (which is
  private to the `biome::rest_api` module) with in-line code in the
  key management routes
* Remove the `get_authorized_user` function, which is no longer used

The resource definitions used by the key management REST API are copied
to the new location rather than moved since there are user endpoints in
the main Biome resource provider that use these resources.

Signed-off-by: Logan Seeley <seeley@bitwise.io>