Stabilize registry-database experimental feature
This commit stabilize the experimental features by
removing the registry-databse features and replacing
it with registry or diesel where appropriate.

Also updates splinterd to for the removal of the features
and updates splinterd to use the database backends
instead. This removes the ability to use a yaml
registry as the local registry.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>