Split orchestrator into builder, runnable, running
This change splits the ServiceOrchestrator into a Builder, a
RunnableServiceOrchestrator (i.e. configured but not started) and a
started ServiceOrchestrator (this is the guts of the original version).

This change allows for the orchestrator to be constructed to a state
that is not started, which can be passed to other components that will
manage its start state.

Likewise, it makes implementing the shutdown handle easy, the state of
the orchestrator is running by its type definition.

The builder defines several defaults for message buffer sizes.

Additionally,

- Display is implemented on the service definition to make reporting
  about service shutdown failures easier
- NewOrchestratorError removes the requirement that errors be send (this
  error is effectively deprecated, replaced with invalid state and
  internal errors).
- The splinter daemon and tests have been updated to use the new builder
  and start function.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>