Duplicate 2PC for admin service
Creates a copy of the existing two-phase commit consensus algorithm that
will be used exclusively by the admin service. This separation will
allow the main 2PC implementation to be altered in ways that would break
compatibility with the admin service without having to update the admin
service.

In the future, this implementation will be removed and the admin service
will be updated to use the updated 2PC implementation.

Signed-off-by: Logan Seeley <seeley@bitwise.io>