Disable custom verifiers list for 2PC
Updates the two-phase commit (2PC) consensus algorithm to no longer
allow the proposal manager to provide a custom list of verifiers for a
proposal. Instead, the verifiers list is now a static list of all peer
nodes and the local node. This means that the coordinator is the same
for all proposals.

This change will allow a node to always know who the coordinator is for
any proposal (since the coordinator is always the same node). This
enables another change that will have only the coordinator creating
proposals, which will be in a future commit.

This change eliminates a known limitation of the 2PC algorithm.

Signed-off-by: Logan Seeley <seeley@bitwise.io>