Replace 2PC state checks with State methods
Replaces most of the state checks in the 2PC consensus algorithm with
use of new methods on the `State` object. This makes the calling code
simpler in a few places. In some cases,
`if let State::EvaluatingProposal(_) = self.state` matches are still
used when the proposal being evaluated needs to be altered.

Signed-off-by: Logan Seeley <seeley@bitwise.io>