Only allow coordinators to create 2PC proposals
Updates the two-phase commit consensus algorithm to only have the
coordinator create new proposals. This change allows us to simplify the
algorithm in a few ways and make some improvements:

* Replace the `proposal_backlog: VecDeque<TwoPhaseProposal>` with
  `proposals_received: HashSet<ProposalId>`, which makes prevention of
  duplicates simpler. We can use a HashSet now because order does not
  matter to non-coordinators: they just verify proposals in the order
  the coordinator requests them. We also just need the proposal ID for
  this.
* Handle `ProposalCreated` and `ProposalsReceived` updates more
  directly, since we know when these are going to occur and on which
  nodes.
* Add some extra checks to the proposal update handler to ensure that
  unexpected messages are logged appropriately.

Because this changes one of the fundamental characteristics of the
algorithm, 2PC now expects the proposal manager to work a bit
differently. Specifically, proposal managers for 2PC will need to ensure
that all proposal data is sent to the coordinator so it can be proposed
by that node. This will require changes to the scabbard service.

Signed-off-by: Logan Seeley <seeley@bitwise.io>