Update ShutdownHandle.wait_for_shutdown to take self
This changes the method to consume the struct that is being
shutdown, thus avoiding leaving the struct in an invalid state.

Due to constraints on traits, updating wait_for_shutdown to
take self also makes it so the trait can no longer be Boxed
and call wait_for_shutdown at the same time. All methods
that returned or took a Boxed<ShutdownHandle> have either been
removed or updated to return the underlying struct directly.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>