Modify openid profile provider for azure photo
Modify the OpenIdProfileProvider to make a call to the microsoft graph
api endpoint for retrieving profile pictures if Azure is being used as
the OpenID provider for auth.

To retrieve the profile picture associated with a microsoft account
an extra call to the microsoft graph api with the auth token is
necessary.

The microsoft graph api /photo/$value endpoint returns the binary data
for the photo. This data is converted to a base64 string and set as
the 'picture' field in the OpenIdProfileResponse struct.

Signed-off-by: Isabel Tomb <tomb@bitwise.io>