Configure Network nodes' registries
This change configures the registry of each node in the network to
include all of the nodes and the admin public key of each. This will
enable all participants to create circuits with each other.

As each node has not yet been configured with a network endpoint, the
endpoint is hard-coded to a valid (but not accessible) endpoint.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>