Add `circuit-purge` feature to services
This change adds the `circuit-purge` feature to the services, including
health and Scabbard. This change also adds the `circuit-purge` feature
guard to the services' `purge` implementation and within the
`ScabbardState` module around the methods used to remove the service
state, used by Scabbard's `purge` implementation.

This change also adds the `scabbard/circuit-purge` and
`health/circuit-purge` dependencies to the `circuit-purge` feature in
splinterd. This allows for the services to be compiled with this feature.

Signed-off-by: Shannyn Telander <telander@bitwise.io>