Remove separate list used for disband proposals
This change removes the separate struct that held `CircuitProposal`s
associated with a `CircuitDisbandRequest` and moves these proposals into
the existing list that holds circuits that had been committed to state
but still need to handle the services.

This also removes the separate handling of the `DisbandedCircuit` message,
as the node now sends a `MemberReady` message instead. The admin service
now inspects the circuit proposal associated with the `circuit_id` sent in
the `MemberReady` message to determine whether the services need to be
initialized as a circuit is created or if the services must be stopped
as a circuit is being disbanded.

Signed-off-by: Shannyn Telander <telander@bitwise.io>