Change method to get reference to Network node
This change alters the `Network`'s `node` method to return a reference
to the node, rather than a mutable reference. This allows the nodes to
be used in the tests directly, rather than having to save values which
was necessary because multiple mutable references can not be used at
once.

Signed-off-by: Shannyn Telander <telander@bitwise.io>