Use reference to nodes in helper functions
This change changes the `commit_2_party_circuit` and
`commit_3_party_circuit` to take references to the nodes themselves,
rather than information pulled from the nodes. These functions are also
updated to create the `node_info` list needed to populate the circuit,
rather than take this as an argument as it can now pull these values
from the nodes directly. This is possible as the Network's `node` method
now returns a reference to the node, rather than a mutable reference
which did not allow for borrowing multiple nodes apart of the network.
This also updates the currently implemented tests to use the information
and methods directly from the node, rather than from saved values.

Signed-off-by: Shannyn Telander <telander@bitwise.io>