Refactor splinterd Dockerfile to fix experimental builds
Compile flags are only passed to dependencies of the crate being
compiled but the workspace Cargo.toml causes all crates to be
recompiled. Because of this, CLI builds will fail when run second
because health and scabbard source files with gated features are
present but the --experimental flag is not being passed.

Signed-off-by: Ryan Beck-Buysse <rbuysse@bitwise.io>