Wait until all nodes have committed the change
This change updates the currently implemented circuit integration tests
to incorporate all nodes in the loop that waits and checks for the change
to be committed. This ensures that all nodes have reached the same state
before moving onto the next part of the test.

This change also updates the comments describing the tests as the
process for validating the change has been committed now includes all
nodes instead of just the nodes receiving the payload from the requester.

Signed-off-by: Shannyn Telander <telander@bitwise.io>