Validate duplicate proposals are rejected
This change adds a section to both the `commit_2_party_circuit` and
`commit_3_party_circuit` functions to submit a duplicate proposal to
create a circuit to the nodes that did not originally submit the circuit
proposal.

This change also adds sections to the `test_3_party_circuit_lifecycle`
and `test_2_party_circuit_lifecycle` to submit a duplicate proposal to
disband a circuit to the nodes that did not originally submit the
disband proposal.

These changes validates that duplicate proposals will be rejected immediately.

Signed-off-by: Shannyn Telander <telander@bitwise.io>