Add scabbard client to Node
This change adds a scabbard client instance to the running Node
instance.  It is configured in a similar manner to the Admin client,
though it uses a builder.  The build error is mapped to an InternalError
as a result.

It requires adding the scabbard/client feature to the node feature.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>