Refactor key loading and signing
This modifies the code to use crate::signing::load_signer instead of the
previous functions, with the exception of some of the registry code
which takes multiple key files.

This fixes a usability bug where most of the commands couldn't load
a key without a private key file being explicitly specified as an
option.

Signed-off-by: Shawn T. Amundson <amundson@bitwise.io>