Add `sabre-sdk` and `transact` dev-dependencies
This change adds the `sabre-sdk` and `transact` crates in the
`dev-dependencies` section of splinterd's Cargo.toml file. This allows
the admin service integration tests to use the Sabre SDK to create a
payload to submit to the Scabbard service. The Transact dependency
allows the integration test to create the `Batch` containing the payload
that is submitted to the Scabbard service.

Signed-off-by: Shannyn Telander <telander@bitwise.io>