Add external registries to testing framework
Add the ability to add a yaml file as an external registry to the
testing framework. If any external registries are added create a unified
registry with the internal registry and external read-only registries.

A yaml registry file can be added by passing the registry file names in
a vector to the `with_external_registries` method when creating a new
Network.

Signed-off-by: Isabel Tomb <tomb@bitwise.io>