Add service txns to circuit integration tests
This change also adds committing service transactions via the nodes'
`Scabbard` client, to validate a service transaction may be committed on
the circuit that has been created. These additions also validate that a
duplicate service transaction is not able to be committed. As well as
validating that service transactions cannot be committed to a disbanded
circuit. This effects the tests in the `circuit_disband` and
`circuit_create` modules.

Signed-off-by: Shannyn Telander <telander@bitwise.io>