Add helper function to create a node's service ID
This function takes a reference to a node and a `circuit_id` used to
create the `ServiceId` struct needed to pass into the `Scabbard` client's
`submit` function. This function is at the top-level mod file as it is
used by the `circuit_create` and `circuit_disband` tests, and will
eventually be used by future `circuit_*` integration tests.

Signed-off-by: Shannyn Telander <telander@bitwise.io>