Remove `circuit-purge` feature from libsplinter
This change removes the experimental `circuit-purge` feature from
libsplinter. This also removes the `splinter/circuit-purge` dependency from
external components, including the CLI, Splinter daemon, and the Health
and Scabbard services. This change also adds `purge` functions to the
Health and Scabbard's `Service` trait implementations, as the
`circuit-purge` feature is still experimental in these components. The
`circuit-purge` feature wiill be stabilized separately in the CLI,
splinter daemon, and the Health and Scabbard services.

Signed-off-by: Shannyn Telander <telander@bitwise.io>