Update 3-party lifecycle to pass
This change updates an `if` statement in the
`test_3_party_circuit_lifecycle` to use `or` instead of `and`, in order
to keep the loop running until all nodes have the updated
`CircuitProposal` containing at least one `VoteRecord` for the disband
proposal. This allows the test to pass, so this change also removes the
`ignore` from this test.

Signed-off-by: Shannyn Telander <telander@bitwise.io>