Implement scabbard signing module similar to splinter's
This replaces crate::key with crate::signing, where crate::signing is
the same as the one in the splinter cli with slight differences due to
CliError definitions.

Also removes the .required(true) from clap, because a default is now
used if not specified.

This is intended to make key-handling the same between the two CLI
commands.

Signed-off-by: Shawn T. Amundson <amundson@bitwise.io>