Add experimental support to splinterd
If "metrics" are enabled, an InfluxDB metrics
record will be initialized to send metrics data
to a running InfluxDB.

This requires providing the name of the database and URL
as well as a username and password for authorization/authentication
with the database. These can be provided with the CLI,
config file or as an environment variables.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>