Update scabbard factory to handle csv service arguments
Before peer_services and admin_keys were treated as json lists.
This commit updates the scabbard factory to check if the
argument string starts with "[" and user json to get values,
otherwise splits the values on ",".

This makes it easier to add arguments to proposals/circuits
without the special formatting.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>