Update propose to take service argument value directly for CLI
Before when adding a service argument with the CLI, the value
was wrapped in a Vec. This is a poor assumption to make and
the CLI should instead set the value as provided. This also
allows for providing multiple options separated by a comma.

To set the argument as before it must be wrapped in [] like this:

--service-arg "*::admin_keys=[\"key\"]"

This also fixes an issues where setting scabbard version with
the CLI was not possible because it would be wrapped in a Vec
when ScabbardFactory expected it would not be.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>