Return admin service event client from Node
This change adds a function `admin_service_event_client`, via the
AdminSubsystem.

Like the REST API variant enums, this commit creates an
AdminServiceEventClientVariant - currently with one option an
implementation backed by the websocket client. The ActixWebClient
variant, in the running case, owns an event reactor, which then will be
shutdown with the admin subsystem.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>