Update `Network` nodes field
Create a `NetworkNode` enum with NetworkNode::Node(Node) and
NetworkNode::RunnableNode(RunnableNode)

Change the `nodes` field of `Network` to be a vector of NetworkNodes
this allows for the vector to hold both the running and stopped nodes
and maintain the original order in which they are generated.

Modify the `add_nodes_with_defaults` and shutdown methods to handle
this change.

Signed-off-by: Isabel Tomb <tomb@bitwise.io>