Update `get_new_access_token` Biome client method
This change updates the `BiomeClient`'s `get_new_access_token` method to
deserialize the response body into a struct, `ClientToken` to allow for
the value of the JSON body to be returned. Previously, deserializing
this response body into a string returned an error, as the `String` type
does not match the response body's structure.

Signed-off-by: Shannyn Telander <telander@bitwise.io>