Update missing protocol id in unpeered payloads
This change fixes an issue where a pending admin payload is waiting on
two or more peers where the peers will have connected, but the pending
payload is never submitted as a proposal via consensus.

The main issue is caused by the fact that the pending payload moves
through two collections during the phases of peering for admin
consensus.

The first phase involves peering, where the connections
authorize each other.  While in this phase, the pending payload is in an
`unpeered_payloads` collection.  Each payload in this collection has
both a list of pending peers and missing protocol ids.  During the first
proposal between three nodes, where no node is peered, both nodes (and
their admin service ids) are included in each of those lists. Once all
the nodes have peered, the pending payload is moved to a
`pending_protocol_payloads` collection.

The second phase involves waiting on the protocol agreement of all the
members.  Each peer must perform a protocol agreement handshake, to
ensure that they can speak via a mutually supported protocol. Once all
the protocols have been agreed to, the payload can be moved to
consensus.

Prior to this fix, if a member node has peered and performed its
protocol agreement exchange before the other unpeered members have
peered, the node's admin service id is never removed from the
`missing_protocol_ids` list for that pending payload. As a result, the
payload never continues on to consensus, due to only the
`pending_protocol_payloads` being evaluated on-protocol-agreement.

Updating the pending payloads in the `unpeered_payloads` collection to
remove any missing protocol ids fixes the issue.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>