Add macro for number of pending batches
This commit includes adding an experimental metrics features
to scabbard and pulling in the libsplinter  no-op macros if
metrics is not enabled. This allows for adding metrics to the
scabbard code without pulling the rust metrics library when
not in use.

Also the scabbard/metrics features was added to the splinterd
metrics feature.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>