Add OAuthConfigBuilder
This commit adds an OAuthConfigBuilder struct in the daemon::oauth
submodule.  This struct produces a runnable OAuth configuration from
optional string values.  It may produce None, if none of the base
required fields are provided.

This builder differs from direct building, as it makes decisions on
whether or not to provide the config, as well as validating the inputs
themselves.

The runnable configuration is converted to an OAuthConfig with the use
of a store factory.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>