Add count methods for proposal and circuits to AdminServiceStore
This commit adds two new methods, count_circuits and count_proposals,
to the AdminServiceStore to return the number of circuits and proposals.
These methods both take the same circuit CircuitPredicates based to the
associated list methods.

These implementations are currently behind an experimental feature
"admin-service-count"

These methods will be used to make adding metrics that keep track
of the number circuits and proposals easier to implement.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>