Add `enable_biome` method to `Node` builder
This change adds a method to the `Node` builder module to enable Biome
within the node. If this method is used, Biome REST endpoints will be
made available to the running instance of the node. This change adds the
necessary plumbing to the `RunnableNode` module to add the Biome
resources to the node's REST API. This also adds the authorization
configuration used by Biome endpoints to the node's REST API.

Signed-off-by: Shannyn Telander <telander@bitwise.io>