Stabilize splinterd oauth
This feature stabilizes the splinterd oauth feature by moving it to the
default set of features, making it a peer of biome-credentials.

The position in default features maintains this feature in all stable
builds, but doesn't not make the feature an optional add on, but still
removable in custom, no-default-features builds.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>